requirements.txt
mitmproxy==10.2.4
playwright==1.42.0
torch==2.2.1
requests
numpy==1.26.4
pillow
--find-links https://github.com/shinkuan/Akagi/releases/expanded_assets/v0.1.1-libriichi
riichi>=0.1.1
tkhtmlview
pyinstaller

# majsoulmax dependencies
protobuf==3.20.1
requests
ruamel.yaml